add service behavior
